Correctly deserialize null values in CRUDs
Previously values stored as null was deserialized as empty strings, etc.
This patch introduce normalization of the deserialization of data read
from the database using shared methods in the superclass.
The patch also change the identities of test data due to the requirements
of UUID v4 complient identities previously introduces in class UUID.

Bug: T267109
Change-Id: I50d50f9bff70c63a0f6354bcea1c2bbda9384d6c