Introduce Persistent API I/O serialization
Serialize and deserialize Persistent instances to/from
associative array middle tier for API I/O.

Includes refactor of the Persistent unit tests.

Also includes introduction of matcher IsJsonSame that matches
the semantics of of two JSON strings, which is added to the
UserDialectEqualsConstraint to avoid failing due to non semantic
differences such as whitespace in the location geojson field.

Bug: T267117
Bug: T268470
Change-Id: Id2061e06ef50096ed57f7006cd1ebf407b26c8cf