Remove use of &$byref from CRUD
Instances of Persistent does not need to be passed byref in order
for the instance passed by the caller to be modified.

The array passed for serialization to database columns needed to
be passed as byref, but the pattern was refactored to remove any
use of byref by creating the array at PersistentCRUD-level and
passing it back as the return value to AbstractCRUD-level for
further processing.

Bug: T266004
Change-Id: Ie0f7401e063402d0313598e9438f9dd8844a4e3d