Add recording functionality to SpecialRecordSpeech
Adds the ability to record speech via microphone, using the
LinguaRecorder library from LinguaLibre
(https://github.com/lingua-libre/LinguaRecorder).

Once the user has given permission to the browser to use the
microphone, they can click the start button to start recording. The
recorder will then listen for a loud enough sound to start recording
and stop when there has been a long enough pause. `stopDuration` was
increased to not stop recording when pausing for a comma. The
parameters have only been tested during development and can be tweaked
further if needed. There is also a split second wait before the
recorder starts listening, to avoid picking up sound from mouse or
keyboard when the user starts the recorder.

Sound data is stored in an <audio> element that allows the user to
review the recording. Making a new recording overwrites the data.

Adds documentation generation with jsdoc.

Bug: T267783
Change-Id: I82a4f17cea3ca16feeceed6bebac80f922a50aa6