Hotfix: Replace ?object typehit for CRUD::read
Replaced with Persistent.

This is an artifact from the very begining of the project when I was
unaware of the meaning of 'object' in PHP, thinking it was any user
defined class, and that the polymorphysism of PHP allowed for overriding
typehints with direct subclasses of the typehinted class type.

I was wrong on both instances.

Change-Id: Ic05966ac71c9758f58276595b8e72e9a849cb5ab