Introduce CRUDContext
Rather than passing down an ILoadBalancer to AbstractRdbmsCRUDs,
we pass down the new class CRUDContext that contains the instance
of ILoadBalancer to be used.

This allows for adding persistency layer contexts for CRUDs that
extend AbstractCRUD. I.e. this makes room for the upcoming CRUD that
store Persistent objects in MediaWiki using Multi Content Revisions.

Bug: T273448
Change-Id: I90a833ae273751a97c10341e24eb2e30948a8a41