CRUD using Multi content review as persistent layer
Introduces AbstractMcrCRUD and implementation RecordingAnnotationsCRUD,
which use Multi Content Revisions in MediaWiki to store implementations
of Persistent as a SlotRecord within a WikiPage.

This will initially be used to store speech recording annotations in a
way that gives us revision history, rollback and summary support on
changes authored by users.

Bug: T272023
Change-Id: I126b47b9a3a72907ad063871d99da068b1f43d39